requirements.txt
openai>=1.0.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
